Change to multi-label classification approach
Previous approach (splitting samples with multiple labels) increased
the number of samples from ~22k to ~55k. Also splitting the samples
caused problems for getting reasonable metrics. The main motivation
for avoiding multi-label classification was to simplify the problem;
as this actually acts the other way, the single-label approach is
dropped.